Fix box render issue and treefarm nbt issue
The box render had an issue where it moved the rendering of everything
down by the player's eyeheight. This happened because I set the
translation of the buffer builder down by the player's height in
`RenderHelper` but did not reset it after rendering. I fixed this by
reseting the translation after rendering in each of the renderhelper's
draw methods.

The tree farm had an issue where it was trying to read the state from
the nbt even when the nbt didn't contain the state. I fixed this by
checking if a key exists in the nbt before querying it's value.