Remove gradle.properties and fix TileEntity saving issues
Removes gradle.properties because it is now handled by the multimod
project setup completely.

Fixes saving issues with TileEntities (markDirty() calls were added
where they should be). Also makes the Tree Farm save as being in the
SCANNING state if it was breaking blocks so nbt isnt flooded with a
bunch of BlockPos tags.